Allow formats for import besides json and xml.
You can configure smuggler to export a format besides json and xml,
but the form for importing data does a manual check on the file
extension for json or xml.

This patch instead checks the file extension against all registered
serializers so that you can load data using a custom serializer.